requirements/dev.txt
# install the prod requirements

-r prod.txt

# for dev
flask-cors
flask-jwt-oidc

# testing
pytest

# Lint and code style
flake8
flake8-blind-except
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
pep8-naming
